pyproject.toml
[project]
name = "02807-project"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.7"
dependencies = [
    "geopandas>=1.1.1",
    "matplotlib>=3.10.7",
    "netcdf4==1.7.3",
    "notebook>=7.4.7",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "scipy==1.16.3",
    "seaborn>=0.13.2",
    "xarray>=2025.10.1",
    "tsfresh>=0.21.1",
    "dtaidistance>=2.3.13",
    "scikit-learn-extra>=0.3.0",
]
